feat: use treesitter_module
