+++ plugin/ts_context_commentstring.vim
@@ -1,0 +1,1 @@
+lua require('ts_context_commentstring').init()
